Local storage update
Optimizing how local storage is changed and moving the parsing logic
to it's own module
