+++ src/scripts/graphElo.js
@@ -8,7 +8,8 @@ import 'chartjs-adapter-moment';
 
 Chart.register(zoomPlugin, LinearScale, PointElement, Tooltip, Legend, TimeScale); 
 
-import { getArchivedGames, getUserName, parseGameNode } from './utils.js';
+import {  getUserName, } from '@/scripts/utils.js';
+import { getArchivedGames, parseGameNode } from '@/scripts/archiveUtils.js';
 
 
 export function graphElo(timeClass="rapid")  {
@@ -26,10 +27,6 @@ export function graphElo(timeClass="rapid")  {
     }
 
 
-   
-
-
-    
     // let timeFormat = "YYYY-MM-DD HH:mm";
     const eloChart = new Chart(ctx, {
         type: 'line',
@@ -142,7 +139,8 @@ export function graphElo(timeClass="rapid")  {
   let allData = [];
 
   for (var i=0; i<archivedGames.length; i++) {
-      let parsedGameNode = parseGameNode(archivedGames[i],uname);
+      // let parsedGameNode = parseGameNode(archivedGames[i],uname);
+      let parsedGameNode = archivedGames[i]
       if (parsedGameNode.timeClass == timeClass){
         let rating = parsedGameNode.userRating;
         let link = parsedGameNode.gameUrl;
