+++ src/scripts/utils.js
@@ -1,3 +1,6 @@
+
+import { getArchivedGames, getParsedArchivedGames } from "@/scripts/archiveUtils.js";
+
 export function clearCharts() {
     const graphs = document.querySelectorAll('.chart');
     graphs.forEach(graph => {
@@ -7,28 +10,12 @@ export function clearCharts() {
 
 export function clearLocalStorage() {
     localStorage.clear();
-    const inlineStorage = document.querySelector('#inlineStorage');
-    const parsedInlineStorage = document.querySelector('#parsedInlineStorage');
+    console.log("local storage cleared")
+    let inlineStorage = document.querySelector('#archivedGames');
     if (inlineStorage) {
         inlineStorage.remove()
+        console.log("inline storage cleared")
     }
-    if (parsedInlineStorage) {
-        parsedInlineStorage.remove()
-    }
-}
-
-export function getArchivedGames() {
-
-    // console.log(window.localStorage.getItem("archivedGames"));
-    if (window.localStorage.getItem("archivedGames") != null) {
-        let archive = window.localStorage.getItem("archivedGames");
-        return JSON.parse(archive);
-    } else {
-        let inlineDiv = document.getElementById("inlineStorage");
-        let archive = inlineDiv.textContent;
-        return JSON.parse(archive);
-    }
-
 }
 
 
@@ -54,128 +41,6 @@ export function utcToHuman(unixTimestamp) {
 
     return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
 
-}
-
-export function parseGameNode(gameNode) {
-    let uname = window.localStorage.getItem("userName");
-
-    let parsedGameNode = {};
-
-    // easy ones
-    parsedGameNode["unixTimeStamp"] = gameNode.end_time;
-    parsedGameNode["timeClass"] = gameNode.time_class;
-    parsedGameNode["gameUrl"] = gameNode.url;
-    parsedGameNode["fen"] = gameNode.fen;
-    parsedGameNode["timeStamp"] = utcToHuman(gameNode.end_time);
-
-    let ogPgn = gameNode.pgn;
-    let stripedPgn = ogPgn.replace(/(\r\n|\r|\n)/g, '');
-    parsedGameNode["pgn"] = stripedPgn;
-
-    parsedGameNode["ogPgn"] = gameNode.pgn;
-
-    // find game color
-    if (gameNode.white.username.toUpperCase() == uname.toUpperCase()) {
-        parsedGameNode["userColor"] = "white";
-        parsedGameNode["result"] = gameNode.white.result;
-        parsedGameNode["opponent"] = gameNode.black.username;
-        parsedGameNode["opponentUrl"] = `https://www.chess.com/member/${gameNode.black.username}`;
-
-        parsedGameNode["opponentRating"] = gameNode.black.rating;
-        parsedGameNode["userRating"] = gameNode.white.rating;
-
-        // eslint-disable-next-line no-prototype-builtins
-        if (gameNode.hasOwnProperty("accuracies")) {
-            parsedGameNode["userAccuracy"] = gameNode.accuracies.white;
-            parsedGameNode["opponentAccuracy"] = gameNode.accuracies.black;
-        } else {
-            parsedGameNode["userAccuracy"] = "";
-            parsedGameNode["opponentAccuracy"] = "";
-        }
-
-    } else {
-        parsedGameNode["userColor"] = "black";
-        parsedGameNode["result"] = gameNode.black.result;
-        parsedGameNode["opponent"] = gameNode.white.username;
-        parsedGameNode["opponentUrl"] = `https://www.chess.com/member/${gameNode.white.username}`;
-
-        parsedGameNode["opponentRating"] = gameNode.white.rating;
-        parsedGameNode["userRating"] = gameNode.black.rating;
-
-        // eslint-disable-next-line no-prototype-builtins
-        if (gameNode.hasOwnProperty("accuracies")) {
-            parsedGameNode["userAccuracy"] = gameNode.accuracies.black;
-            parsedGameNode["opponentAccuracy"] = gameNode.accuracies.white;
-        } else {
-            parsedGameNode["userAccuracy"] = "";
-            parsedGameNode["opponentAccuracy"] = "";
-        }
-
-    }
-
-    // find out how you won
-    parsedGameNode["wonBy"] = "";
-    if ((parsedGameNode.userColor == "white") && (parsedGameNode.result == "win")) {
-        parsedGameNode["wonBy"] = gameNode.black.result;
-    }
-
-    if ((parsedGameNode.userColor == "black") && (parsedGameNode.result == "win")) {
-        parsedGameNode["wonBy"] = gameNode.white.result;
-    }
-
-    // pgn parsing
-    let pgn = gameNode.pgn.split('\n');
-    // eslint-disable-next-line no-useless-escape
-    parsedGameNode["date"] = pgn[2].replace(/\\|\[|\]|\"|Date/g, '');
-
-    // find opening url. The fact we have to do this means something is broken
-    let openingUrl = "";
-    for (let i = 0; i < pgn.length; i++) {
-        if (pgn[i].startsWith("[ECOUrl")) {
-            openingUrl = pgn[i];
-            break;
-        }
-    }
-    // eslint-disable-next-line no-useless-escape
-    parsedGameNode["openingUrl"] = openingUrl.replace(/\\|\[|\]|\"|ECOUrl/g, '');
-    // eslint-disable-next-line no-useless-escape
-    let tmp_opening = openingUrl.replace(/\\|\[|\]|\"|ECOUrl|https:\/\/www.chess.com\/openings\//g, '');
-    parsedGameNode["opening"] = tmp_opening.replace(/-/g, " ");
-    parsedGameNode["opening"] = parsedGameNode["opening"].trimStart();
-
-    let mainLine = parsedGameNode["opening"].match(/^(\D*)(?=\d)/);
-    if (mainLine) {
-        parsedGameNode["mainLineOpening"] = mainLine[1];
-    } else {
-        parsedGameNode["mainLineOpening"] = parsedGameNode["opening"];
-    }
-
-    // eslint-disable-next-line no-useless-escape
-    parsedGameNode["startTime"] = pgn[17].replace(/\s|\[StartTime|\]|\"/g, '');
-    // eslint-disable-next-line no-useless-escape
-    parsedGameNode["endTime"] = pgn[19].replace(/\s|\[EndTime|\]|\"/g, '');
-
-    // ugly  
-    parsedGameNode["gameId"] = parsedGameNode["gameUrl"].match(/(live|daily)\/(.*)$/)[2];
-
-
-    // main line openings 
-
-    return parsedGameNode;
-}
-
-export function verifyLiveChess(gameNode) {
-
-    try {
-        if ((gameNode.rules == "chess" || gameNode.rules == "chess960")) {
-            return true;
-        } else {
-            return false;
-        }
-
-    } catch (error) {
-        return false
-    }
 }
 
 
@@ -231,8 +96,6 @@ export function isTop90Percentile(number, numbers) {
 }
 
 
-
-
 export function getWinsAndLossesByOpenings(timeClass, opening, parsedArchivedGames) {
     let winCount = 0;
     let lossCount = 0;
@@ -273,42 +136,7 @@ export function getWinsAndLossesByOpenings(timeClass, opening, parsedArchivedGam
 }
 
 
-export function parseAndSaveArchivedGames() {
-    let parsedArchivedGames = []
-    let archivedGames = getArchivedGames()
-    for (let i = 0; i < archivedGames.length; i++) {
-        let gameNode = parseGameNode(archivedGames[i])
-        parsedArchivedGames.push(gameNode)
 
-    }
-    console.log(`parsedArchivedGames: ${parsedArchivedGames.length}`)
-    try {
-        window.localStorage.setItem("parsedArchiveGames", JSON.stringify(parsedArchivedGames));
-        console.log("parsedArchiveGames saved to local storage")
-    } catch (err) {
-        let inlineStorage = document.createElement("div");
-        let appDiv = document.getElementById("app");
-        inlineStorage.setAttribute("id", "parsedInlineStorage");
-        inlineStorage.setAttribute("hidden", "hidden");
-        inlineStorage.textContent = JSON.stringify(parsedArchivedGames);
-        appDiv.appendChild(inlineStorage);
-        console.log("parsedArchiveGames saved to inline storage")
-    }
-
-}
-
-export function getParsedArchivedGames() {
-
-    if (window.localStorage.getItem("parsedArchiveGames") != null) {
-        let parsedArchive = window.localStorage.getItem("parsedArchiveGames");
-        return JSON.parse(parsedArchive);
-    } else {
-        let inlineDiv = document.getElementById("parsedInlineStorage");
-        let parsedArchive = inlineDiv.textContent;
-        return JSON.parse(parsedArchive);
-    }
-
-}
 
 export function saveOpeningsData() {
     // let timeClasses = ["all", "bullet", "blitz", "rapid", "daily"]
@@ -333,7 +161,7 @@ export function saveOpeningsData() {
 
 export function calculateOpening(timeClass) {
 
-    let parsedArchivedGames = getParsedArchivedGames();
+    let parsedArchivedGames = getArchivedGames();
     // let uname = getUserName();
 
     const openingData = {};
@@ -424,15 +252,7 @@ export async function fetchUserStats(userName){
     let playerStatsUrl = `https://api.chess.com/pub/player/${userName}/stats`;
     let playerStatsRes = await fetch(playerStatsUrl);
     return playerStatsRes;
-    // let playerStats = await playerStatsRes.json();
 
-    // if (playerStatsRes.status != 200 ) {
-    //     alert("User Not found");
-    //     return "error";
-    //     // window.localStorage.setItem("playerStats", JSON.stringify(playerStats));
-    // } else {
-    //     return playerStats;
-    // }
 
 }
 
@@ -442,15 +262,4 @@ export async function fetchArchiveUrls(userName) {
     let response = await fetch(archiveUrl);
 
     return response; 
-    // let archiveMonths = await response.json();
-
-    // if (response.status != 200 ) {
-    //     alert("User Not found");
-    //     return "error";
-    // }
-    // else {
-
-    // return archiveMonths.archives
-    // }
-
 }
